Provide a clear log message when a cluster can't be handled
When a cluster is missing its id.k8s.io cluster claim, it can't be
handled. There was no indication that this was happening except a debug
log message that is not displayed by default. This commit makes the log
message clearer and changes it to a regular info log.

Relates:
https://issues.redhat.com/browse/ACM-2625

Signed-off-by: mprahl <mprahl@users.noreply.github.com>